+++ python/examples/01_the_switch_and_the_light/switch.py
@@ -1,10 +1,13 @@
 import atexit
+import ssl
 import sys
 import time
 import RPi.GPIO as GPIO
-from client import Client
 from threading import Thread
 from time import sleep
+
+sys.path.append('../../src')
+from client import Client
 
 GPIO04 = 4
 DESIRED_STATE_KEY = 'cns/network/nodes/sri4FZUq2V7S4ik2PrG4pj/contexts/kMqdHs8ZcskdkCvf1VpfSZ/provider/padi.button/connections/geizaJngWyA1AL3Nhn5dzD/properties/sState'
@@ -14,7 +17,10 @@ GPIO.setmode(GPIO.BCM)
 GPIO.setup(GPIO04, GPIO.IN)
 
 # Connect to Arete control plane
-client = Client.connect('wss://dashboard.test.cns.dev:443')
+ssl_context = ssl.SSLContext(protocol=ssl.PROTOCOL_TLS_CLIENT)
+ssl_context.check_hostname = False
+ssl_context.verify_mode = ssl.CERT_NONE
+client = Client.connect('wss://dashboard.test.cns.dev:443', ssl=ssl_context)
 sys.stderr.write('Connected to Arete control plane\n')
 
 # Read initial switch state, and sync it with Arete