+++ python/src/system_id.py
@@ -5,7 +5,7 @@ LINUX_MODEL_FILENAME = '/sys/firmware/devicetree/base/model'
 LINUX_SERIAL_NUMBER_FILENAME = '/sys/firmware/devicetree/base/serial-number'
 
 def get_system_id():
-    if os.path.is_file(LINUX_MODEL_FILENAME) and os.path.is_file(LINUX_SERIAL_NUMBER_FILENAME):
+    if os.path.isfile(LINUX_MODEL_FILENAME) and os.path.isfile(LINUX_SERIAL_NUMBER_FILENAME):
         return get_system_id_linux()
     else:
         raise Exception('Unable to detect System ID on this platform')
@@ -17,5 +17,5 @@ def get_system_id_linux():
         serial_number = file.read()
     model_plus_serial_number = f"{model}:{serial_number}"
 
-    id = uuid.uuidv5(uuid.NAMESPACE_OID, model_plus_serial_number)
+    id = uuid.uuid5(uuid.NAMESPACE_OID, model_plus_serial_number)
     return id